consumers/mongo_writer/requirements.txt
kafka-python
pymongo
requests
python-dotenv
